1. modify process of invalid link 2. develop docs

+++ src/util/file_util.py
@@ -58,6 +58,9 @@ def get_image_to_target(link: str, from_filepath: str, target_foldpath: str) -> 
     if str_util.is_url(link):
         net_util.down_image(link, os.path.join(target_foldpath, name))
     else:
+        if os.path.exists(link) is False:
+            print("该路径不存在: ", link)
+            return name
         shutil.copyfile(link, os.path.join(target_foldpath, name))
 
     return name
